+++ crates/gpx/README.md
@@ -27,17 +27,24 @@ gpx predict --help
 `gpx fit` takes training input as a required positional argument.
 `gpx fit` uses the last `N` columns as outputs via `--outputs N` (default `1`) and
 trains one surrogate model per output column.
+`gpx fit` supports expert-model configuration options:
+- `--regression-spec <constant|linear|quadratic|all>` (default `constant`)
+- `--correlation-spec <squared-exponential|absolute-exponential|matern32|matern52|all>` (default `squared-exponential`)
+- `--kpls-dim <N>` (optional)
+- `--n-clusters <N>` (default `1`, must be >= `1`)
+- `--recombination <hard|smooth>` (default `smooth`)
+- `--smooth-factor <F>` (optional, valid only when recombination is `smooth`)
 `gpx fit` writes to `surrogate_model.gpx` by default; use `-o/--output` to change it.
 `gpx fit` auto-detects model output format from output file extension:
 - `.json` -> `json`
 - otherwise -> `binary`
-`gpx qa` and `gpx spec` use all surrogates by default; use `-i/--model-index`
+`gpx qa` and `gpx spec` use all surrogates by default; use `-m/--model-index`
 to target a single surrogate.
 `gpx predict` takes input as a required positional argument.
 `gpx predict` uses `surrogate_model.gpx` by default; use `--model` to change it.
 By default `gpx predict` uses all surrogate models from the model file and writes
 one output file containing inputs followed by predicted outputs (training-like layout).
-Use `-i/--model-index` to predict a single output model.
+Use `-m/--model-index` to predict a single output model.
 `gpx predict` auto-detects input format from input file extension:
 - `.npy` -> `npy`
 - otherwise -> `csv`
